+++ backend/spec/tests/Users.spec.ts
@@ -2,7 +2,7 @@ import supertest from "supertest";
 import StatusCodes from "http-status-codes";
 import { SuperTest, Test } from "supertest";
 
-import app from "server";
+import app from "app";
 import UserDao from "daos/User/UserDao.mock";
 import User, { IUser } from "entities/User";
 import { paramMissingError } from "shared/constants";
